fix: use client side for the room detail page
